Minify JS; load d3 modules just once
Relates to #46
