+++ assets/js/data-viz/regionAster.js
@@ -1,7 +1,11 @@
-const d3 = Object.assign(
-  {},
-  require("d3-selection"),
-);
+let d3 = window.d3;
+
+// if (!d3) {
+//   d3 = Object.assign(
+//     {},
+//     require("d3-selection"),
+//   );
+// }
 
 import getData from "./data.js"
 import aster from "./aster.js"
