+++ assets/js/data-viz/regionTooltip.js
@@ -1,7 +1,11 @@
-const d3 = Object.assign(
-  {},
-  require("d3-selection")
-);
+let d3 = window.d3;
+
+// if (!d3) {
+//   d3 = Object.assign(
+//     {},
+//     require("d3-selection")
+//   );
+// }
 
 /**
  * The classes to add to the various HTML elements that are combined to create a tooltip.
