+++ layouts/_default/baseof.html
@@ -20,7 +20,10 @@
       
       {{- $jsParams := dict "breakpoint" 900 "dataConfig" $dataConfig -}}
 
-      {{- $script := . | js.Build (dict "params" $jsParams) -}}
+      {{- $script := . | js.Build (dict "params" $jsParams "minify" true) -}}
+
+      {{ $sizeInBytes := len $script.Content }}
+      
       <script src="{{ $script.RelPermalink }}" async></script>
     {{- end -}}
